+++ src/vtex/getSkuById/index.ts
@@ -1,0 +1,3 @@
+import { getSkuById } from './getSkuById';
+
+export default getSkuById;