+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added automation tools to handle documentation.
+- [vtex] Added `getSkuById` method.
 - [vtex] Added `getShelfByCollection` method.
 - [vtex] Added `getShelfById` method.
 - [vtex] Added `getShelfBySku` method.